Refactor feature flags
