+++ plugins/utils/unified-toc.js
@@ -1,19 +1,25 @@
 const { load } = require('js-toml');
+const fs = require('fs/promises');
 
 const createTocNodes = async ({ createNode, createNodeId, createContentDigest }) => {
   // Get all MongoDB products for the sidenav
 
-  createNode({
-    children: [],
-    id: createNodeId('toc'),
-    internal: {
-      contentDigest: createContentDigest(product),
-      type: 'Product',
-    },
-    parent: null,
-    title: product.title,
-    url: product.baseUrl + product.slug,
-  });
+  try {
+    const tomlContents = (await fs.readFile(`${process.cwd()}/toc.toml`)).toString();
+    const toc = load(tomlContents);
+
+    createNode({
+      tocTree: toc,
+      id: createNodeId('toc'),
+      internal: {
+        contentDigest: createContentDigest(toc),
+        type: 'TOC',
+      },
+      parent: null,
+    });
+  } catch (e) {
+    console.error('error occurred when reading the toc.toml', e);
+  }
 };
 
 module.exports = {
