+++ src/utils/feature-flags.js
@@ -1,0 +1,5 @@
+export function getFeatureFlags() {
+  return {
+    isUnifiedToc: Boolean(process.env.GATSBY_USE_UNIFIED_TOC),
+  };
+}
